pyproject.toml
[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "A comprehensive data analysis automation solution for collecting, processing, and analyzing marketing data from multiple sources"
readme = "README.md"
requires-python = ">=3.9"
dependencies = []

[dependency-groups]
dev = [
    # Core data processing
    "pandas>=2.2.3",
    "numpy>=2.0.2",
    "scipy>=1.13.1",
    
    # Google API clients (Enhanced)
    "google-analytics-data>=0.18.18",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=0.4.6",
    "gspread>=5.0.0",
    
    # Configuration and environment
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    
    # Database
    "mysql-connector-python>=9.3.0",
    "sqlalchemy>=2.0.41",
    
    # Visualization
    "matplotlib>=3.9.4",
    "plotly>=6.1.2",
    "seaborn>=0.13.2",
    
    # Machine learning
    "scikit-learn>=1.6.1",
    "statsmodels>=0.14.4",
    
    # Social media APIs
    "facebook-business>=22.0.5",
    
    # Utilities
    "requests>=2.32.3",
    "pytz>=2025.2",
    "tqdm>=4.67.1",
    
    # Testing (optional)
    "pytest>=6.2.5",
]
